+++ port/dto/database_comment.go
@@ -4,7 +4,7 @@ import "time"
 
 type DatabaseComment struct {
 	ID        int       `db:"id"`
-	AuthorID  int       `db:"author_id"`
+	AuthorID  int64     `db:"author_id"`
 	SongID    int       `db:"song_id"`
 	Title     string    `db:"title"`
 	Content   string    `db:"content"`
